moved d2loader into d2common (I dont think it belonged in d2core)
