+++ d2common/d2loader/asset/asset.go
@@ -3,7 +3,7 @@ package asset
 import (
 	"io"
 
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2loader/asset/types"
+	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2loader/asset/types"
 )
 
 // Asset represents a game asset. It has a type, an asset source, a sub-path (within the
