+++ d2common/d2loader/asset/source.go
@@ -1,7 +1,7 @@
 package asset
 
 import (
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2loader/asset/types"
+	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2loader/asset/types"
 )
 
 // Source is an abstraction for something that can load and list assets
