+++ d2common/d2loader/asset/types/asset_types.go
