+++ d2common/d2loader/asset/types/doc.go
