+++ d2common/d2loader/filesystem/asset.go
@@ -3,8 +3,8 @@ package filesystem
 import (
 	"os"
 
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2loader/asset"
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2loader/asset/types"
+	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2loader/asset"
+	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2loader/asset/types"
 )
 
 // static check that Asset implements Asset
