+++ d2common/d2loader/loader.go
@@ -7,10 +7,10 @@ import (
 	"path/filepath"
 
 	"github.com/OpenDiablo2/OpenDiablo2/d2common"
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2loader/asset"
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2loader/asset/types"
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2loader/filesystem"
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2loader/mpq"
+	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2loader/asset"
+	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2loader/asset/types"
+	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2loader/filesystem"
+	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2loader/mpq"
 )
 
 const (
