+++ d2common/d2loader/loader_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2loader/asset"
+	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2loader/asset"
 )
 
 const (
