+++ d2common/d2loader/testdata/A/common.txt
