+++ d2common/d2loader/testdata/A/exclusive_a.txt
