+++ d2common/d2loader/testdata/B/common.txt
