+++ d2common/d2loader/testdata/B/exclusive_b.txt
