+++ d2common/d2loader/testdata/D.mpq
